requirements.txt
astroid==2.3.3
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
colorama==0.4.3
defusedxml==0.6.0
Django==2.2.3
django-allauth==0.39.1
django-braces==1.13.0
django-cors-headers==3.1.0
django-crispy-forms==1.7.2
django-filter==2.2.0
django-markdown-deux==1.0.5
django-oauth-toolkit==1.2.0
django-pagedown==1.0.6
django-rest-auth==0.9.5
django-webpack-loader==0.6.0
djangorestframework==3.10.1
djangorestframework-jwt==1.11.0
google-auth==1.6.3
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
Markdown==3.1.1
markdown2==2.3.8
oauthlib==3.0.2
Pillow==7.2.0
pyasn1==0.4.6
pyasn1-modules==0.2.6
PyJWT==1.7.1
pylint==2.4.4
python3-openid==3.1.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
wincertstore==0.2
wrapt==1.11.2
